*chacha: Scope HChaCha20 only within crate and create subkey_and_nonce generation function to reduce code duplication.
